requirements.txt
streamlit
groq
gTTS
feedparser
python-dotenv
beautifulsoup4
requests
firebase-admin
